requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.0.2
boto3==1.34.0
moviepy==1.0.3
Pillow==10.1.0
requests==2.31.0
beautifulsoup4==4.12.2
numpy==1.24.3
opencv-python==4.8.1.78
pydub==0.25.1
librosa==0.10.1
python-dotenv==1.0.0
aiofiles==23.1.0
httpx>=0.25.0
jinja2==3.1.2
lxml==4.9.2
edge-tts==6.1.10
pyttsx3==2.90 